Rolling back to the previous version of hibernate due to caGRID incompatibilities with caDSR (but keeping the other updated jars).
GForge Tracker #29762
Peer reviewed by Ngoc

+++ caintegrator2-war/src/gov/nih/nci/caintegrator2/data/CaIntegrator2Dao.java
@@ -410,7 +410,7 @@ public interface CaIntegrator2Dao {
      * @param imageSeries to retrieve number of images.
      * @return number of images for source.
      */
-    Long retrieveNumberImages(Collection<ImageSeries> imageSeries);
+    int retrieveNumberImages(Collection<ImageSeries> imageSeries);
     
     /**
      * For a given genomic source, retrieves all distinct platforms.
